lots of catalog changes
